backend check for user is form user session for delete/chage of diary notes, likes added

+++ backend/routes/api/userDiaryRouter.js
@@ -18,8 +18,12 @@ module.exports = userDiaryRouter
     try {
       const { id } = req.params;
       const { userId } = req.body;
-      await UserDiary.destroy({ where: { id, userId } });
-      res.json({ id });
+      if (userId === req.session.user.id) {
+        await UserDiary.destroy({ where: { id, userId } });
+        res.json({ id });
+      } else {
+        res.json({ status: false });
+      }
     } catch (error) {
       res.status(500).send(`${error.message}`);
     }
@@ -35,19 +39,21 @@ module.exports = userDiaryRouter
   })
   .put('/note', async (req, res) => {
     try {
-      const {
-        userId, noteId, changeInputTitle, changeInputContent,
-      } = req.body;
-      const note = await UserDiary.findOne({
-        where: {
-          userId,
-          id: noteId,
-        },
-      });
-      note.title = changeInputTitle;
-      note.content = changeInputContent;
-      await note.save();
-      res.json(note.dataValues);
+      const { userId, noteId, changeInputTitle, changeInputContent } = req.body;
+      if (userId === req.session.user.id) {
+        const note = await UserDiary.findOne({
+          where: {
+            userId,
+            id: noteId,
+          },
+        });
+        note.title = changeInputTitle;
+        note.content = changeInputContent;
+        await note.save();
+        res.json(note.dataValues);
+      } else {
+        res.json({ status: false });
+      }
     } catch (error) {
       res.status(500).send(`${error.message}`);
     }
